extracted look handelr
